requirements.txt
google-genai>=1.0.0
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.35.0
gunicorn>=22.0.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.0
